Separate ANSI parsing from HTML rendering
