+++ lib/livebook_web/live/output/text_component.ex
@@ -12,7 +12,7 @@ defmodule LivebookWeb.Output.TextComponent do
       <%# Add a newline to each element, so that multiple lines can be copied properly %>
       <div data-template class="hidden"
         id={"virtualized-text-#{@id}-template"}
-        ><%= for line <- ansi_to_html_lines(@content) do %><div><%= [line, "\n"] %></div><% end %></div>
+        ><%= for line <- ansi_string_to_html_lines(@content) do %><div><%= [line, "\n"] %></div><% end %></div>
       <div data-content class="overflow-auto whitespace-pre font-editor text-gray-500 tiny-scrollbar"
         id={"virtualized-text-#{@id}-content"}
         phx-update="ignore"></div>
