+++ test/livebook_web/helpers_test.exs
@@ -3,16 +3,6 @@ defmodule LivebookWeb.HelpersTest do
 
   alias LivebookWeb.Helpers
 
-  describe "ansi_to_html_lines/1" do
-    test "puts every line in its own tag" do
-      assert [
-               {:safe, ~s{<span style="color: var(--ansi-color-blue);">smiley</span>}},
-               {:safe, ~s{<span style="color: var(--ansi-color-blue);">cat</span>}}
-             ] ==
-               Helpers.ansi_to_html_lines("\e[34msmiley\ncat\e[0m")
-    end
-  end
-
   describe "names_to_html_ids/1" do
     test "title case" do
       assert(Helpers.names_to_html_ids(["Title of a Section"]) == ["title-of-a-section"])
